fl-server/requirements.txt
Flask
gunicorn
Werkzeug
pytest
torch
torchvision
onnx
onnxruntime-training-cpu